added role to user for admin and user
differenciate our users between admin and plain old users with a role
field in the database. Following some initial advice from

https://naturaily.com/blog/ruby-on-rails-enum

We chose going down the simple path of an integer in the db and
an enum in ActiveRecord. Secure our admin by checking if they
have an admin role.

Co-authored-by: Selena Small <selenawiththetattoo@gmail.com>