added match and play
The play of a game is a match which will be made up of many plays. This
is to be used for web socket communication for real time palys as per
blog
  https://medium.com/@dakota.lillie/using-action-cable-with-react-c37df065f296

The assumption is that play comms can be text but may make more sense to
be json in future?
  bundle exec rails generate model Match title
  bundle exec rails generate Play text match:references